Move dice funcs to file
